requirements.txt
primer3-py==2.0.3
deap==1.4.1
absl-py==2.1.0
astunparse==1.6.3
biopython==1.83
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
dna-features-viewer==3.1.3
dnachisel[reports]==3.2.11
docopt==0.6.2
et-xmlfile==1.1.0
flametree==0.1.12
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.5.4
geneblocks==1.2.3
goldenhinges==0.2.3
google-pasta==0.2.0
grpcio==1.63.0
h5py==3.11.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
tf.keras
kiwisolver==1.4.5
Levenshtein==0.25.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
networkx==3.2.1
Numberjack==1.2.0
numpy==1.26.4
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
proglog==0.1.10
protobuf==4.25.3
Pygments==2.18.0
pyparsing==3.1.2
python-codon-tables==0.1.12
python-dateutil==2.9.0.post0
python-Levenshtein==0.25.1
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.0
requests==2.31.0
rich==13.7.1
six==1.16.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.18.2
WeasyPrint==52.5

